requirements.txt
fastapi
matplotlib
numpy
pandas
Pillow
pydantic
pytest
python-dotenv
PyYAML
seaborn
tensorflow
scikit-learn
opencv-python
uvicorn
httpx